Add start of watch package
This commit adds publish/subscribe code that will form the basis of a
watch package. It corrects some problems with the publish/subscribe
implementation used by the PoC. Events are guaranteed not to be missed,
and guaranteed to be delivered in the same order they were published.

This implementation is a bit mutex-heavy, but that should only matter if
many subscribers watch the same queue. If this use case ends up being
important, the dequeueing process can be changed to use finer-grained
locking.

The next commit will add a public interface to create and destroy
watchers. This will probably be more strongly-typed than this internal
code.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>