api: define declarative network api
Similar to Jobs and Tasks, Networks are first-class declarative
components in the architecture. This change uses libnetwork as
inspiration to define an API in this vein. To pair with network
definition, we allow jobs to specify a NetworkAttachment as part of the
specification. The parameters are minimal at this time but this should
be sufficient.

Signed-off-by: Stephen J Day <stephen.day@docker.com>