Add the ability to "fork" a state
This copies the existing state to a supplied empty Store, and returns a
watcher that will receive any changes from this baseline.

The caller which requested the forked state can call Apply for each Item
it receives on the channel to keep the forked state in sync with the
original one.

This provides the ability to have goroutine-local copies of the state,
to avoid making constant copies of a shared global state or locking that
global state.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>