identity: provide nice identifier generation
This package provides the first parts of the identity package. The first
main component provides human-compatible random identifier generation.
Unlike UUIDs, they provide the full 128 bits of entropy while encoded in
a format that saves 11 bytes. Generated identifiers are also case
insensitive.

Rough parameterization is provided as notes in case we want to tune the
entropy. Identifiers that are under the maximum number of bytes for a
base36, 128-bit number (25 bytes) are padded out with leading zeros.

Signed-off-by: Stephen J Day <stephen.day@docker.com>