Refactor Store interface
This makes the Store interface transactional and splits it up into
smaller interfaces.

The Store interface now has methods to begin read transactions and full
read/write transactions. These transactions have Nodes, Tasks, and Jobs
functions that provide either a readable or read/writable view of the
respective data type. Transactions are protected from interfering with
each other (using a lock, in the current memory store implementation).

Nodes, Tasks, and Jobs return separate interfaces to allow type safe
access. I also experimented with having them return a shared interface,
and found this was clunkier to work with - but it's another possible
model.

Fork has been renamed to Snapshot and now uses a StoreCopier interface
to split out the snapshotting functionality to the destination of the
data, by passing in a read transaction. This allows the destination of
the snapshot to implement a much simpler interface (it only needs a
CopyFrom method).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>