api: ensure declaration of meta and ID are consistent
We address several consistency issues in this change along the
declaration of meta and identifier fields. Meta is now only declared by
user controlled "Spec" types. Identifiers are now declared with proper
Go syntax.

Several TODO statements have been addressed. The biggest change is
moving Source to the top-level and adding a "resolved" source to the
task specification. This allows us to instruct an agent to run an exact
version of the image. The usage will become clearer as we integrate the
runtime.

Signed-off-by: Stephen J Day <stephen.day@docker.com>