Convert planner to use local snapshot of data from store
The planner maintains its own state based on a snapshot of the master
store. It keeps its own indices that make sense for its access patterns.
The local state is kept up to date with the watch channel returned by
Snapshot. Since the local state does not need to support concurrent
access, there is no locking necessary.

Add a unit test for the planner.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>