Change memory store transactions to track in-progress changes
This allows read transactions to happen while an update is in progress,
because update transactions don't mutate the store's internal data
structures until commit time.

Currently, there is a limit of one concurrent update transaction, to
make sure that updates don't conflict with each other, but this could be
changed.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>