agent: implement weighted manager selection
To support connection balancing in an agent, we define an
observation-based, weight random selection. It is integrated with a
grpc.Picker to provide seamless agent reconnection during RPC failure.

The algorithm uses weighted random selection to choose managers such
that the higher the weight, the more likely the manager is to get
selected. We coupled with an exponentially weight moving average to
smooth out observations under flapping, albeit heavily weighted towards
recent values.

Signed-off-by: Stephen J Day <stephen.day@docker.com>