api: rename agent server dispatcher
After working with the agent API, it was found that the code was less
confusing when calling it Dispatcher. Better to make this change early
on, when we have fewer references to it. We also took some time to
document RPC methods and types and order them per agent registration
flow.

Several fields have been slightly changed or modified to support a
reliable registration flow.

Signed-off-by: Stephen J Day <stephen.day@docker.com>