dispatcher: provide session id support
To support clean and race-free agent registration, we now have support
for a session identifier. The identifier is provided on agent
registration to the dispatcher and should be used with other dispatcher
calls. This allows the dispatcher to better control which set of RPC
calls belong to a given post-Register flow. SessionID is also a great
place to launch an authentication context, albeit that needs more
thought.

Signed-off-by: Stephen J Day <stephen.day@docker.com>