api: move NodeID from NodeSpec to Node
We have no plans to allow nodes to declare their own identifiers as part
of the node specification. In all cases, NodeID will be authenticated or
assigned to the node via the manager. Before this becomes a problem, we
are moving NodeID to Node, where it belongs. Nodes still state their
NodeID, as they may see, as part of a `RegisterRequest`, but dispatchers
must authenticate it.

Signed-off-by: Stephen J Day <stephen.day@docker.com>