api: define TaskSpec to control execution
Up until this point, discussion on JobSpec and TaskSpec has been
academic. The goal was to have some set of common fields that would flow
from Job to Task. After implementing a partial agent, it was found that
isolating the cluster-level definitions from the agent-level definitions
is the correct approach. We accomplish this by adding a template to
JobSpec which defines the core task configuration. This field is copied
directly into child tasks of a job.

TaskSpec only contains a single union field at this time. The intent is
to have several different "runtimes" that can be defined on task,
allowing the dispatch and execution of different targets, depending on
requirements. The only available runtime at this point is docker,
defined by ContainerSpec. With this, we omit the concept of a source and
bind the runtime context to the runtime, meaning that each runtime must
have the ability to fetch their own context.

Signed-off-by: Stephen J Day <stephen.day@docker.com>