agent/dockerexec: define and implement task runner
We take our first step into task execution by defining and implementing
the Runner interface. A Runner is a blocking interface focused purely on
task setup and execution. It is implemented against the engine api.
Fully mocked unit tests cover most operations and a limited integration
test will create, run and destroy tasks on the local docker instance, if
given a parameter.

The next step from here is implement a task controller to manager and
report on Runner state. Task state will be inferred by the behavior of
calls to the runner in the main task controller goroutine.

Signed-off-by: Stephen J Day <stephen.day@docker.com>