Only run unit tests once in CI
The unit tests were first being run as part of the "all" makefile
target, and then a second time for "coverage". We only need to run
"coverage".

Introduce a "ci" target in the makefile that builds the appropriate set
of targets (equavelent to "all",  minus "test", plus "checkprotos" and
"coverage").

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>