api: task "Accepted" state and Go idiomatic enum
To ensure we can detect service time between a node receiving a task and
preparing of the runtime environment, an ACCEPTED state has been added.
Once a node has registered the task locally, the ACCEPTED state is
reported to the manager. The main benefit to measure the time between
ACCEPTED and PREPARING to detect overloaded nodes.

The TaskStatus.Message field is now called Err. We may want to provide
more structure there, but this is to clarify that it is not a general
message field. Ideally, this field could serialize arbitrary errors, but
we'd need to do some code generation tricks to acheive this.

We also now generate idiomatic enum definitions.

Signed-off-by: Stephen J Day <stephen.day@docker.com>