agent: implement run and report flow
We define a generic reporting interface to propagate task status based
on the standard run flow. The Run function allows cancelling based on
context or failed task reporting. Conditions signaled by the runner
ensure that we can move forward to the correct call and report the
state, even if state data is completely lost.

The next step is to integrate this will a task controller to ensure this
is managed correctly for every task.

Signed-off-by: Stephen J Day <stephen.day@docker.com>