agent: full task controller flow
We have finally come to a place where the entire task controller flow is
wired to docker. From the command line, we can create jobs on a docker
instance, monitor their status and clean them up. We have full state
reporting and decent err reporting and handling.

A critical separation involved removing the session management from the
agent. It ensures the agent code is purely focused on state and
responding to external signals. As we move from here, we can likely find
abstractions that encapsulate this separation.

There is still much to be desired. Notes are provided within to move the
design in the correct direction. As we break down more area for unit
testing, we will see better componentization.

Signed-off-by: Stephen J Day <stephen.day@docker.com>