agent: refactor to comply with gocyclo check
Normally, it would better to rebase changes based on automated checks
into existing commits. In this case, it seems interesting keep a history
of the changes made based on the check. Is the code clearer? Some may
say so, but it may have been more valuable to spend time on concerted
refactoring work or unit testing. Hopefully, this visibility can help us
make a decision on our tools.

Signed-off-by: Stephen J Day <stephen.day@docker.com>