Optimize scheduler: use a heap to keep track of nodes
This avoids an expensive redundant loop to find the most
lightly-scheduled node that used to happen for every scheduling
decision.

The nodeHeap implementation keeps track of the indices in the heap by
node ID, so externally originating node changes and task changes can be
handled without rebuilding the heap from scratch or doing expensive
searches.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>