Check sequence information on updates to store
Add a Seq field to each type of store object. This stores the raft index
when it was last updated. On updates, check that the Seq field in the
argument matches the information in the store. If it does not, there was
another update since this object was retrieved, and the update should be
rejected.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>