namespaces: experiment with resolution
The enclosed changeset includes a namespace context implementation with
search spaces and configurable resolution. We unit test this at a
production-ready level.

Also included is a simulation of namespace resolution in an environment
with namespaces and search resolution. Unresolved references are left to
the cluster for resolution against namespaces.

Notice that with very few lines of code, we get extremely interesting
behavior. The following is the output of the experiment:

```
namespace ns1
    www == www.ns1
	frontend == frontend.ns3
	backend == backend.ns3
    redis == redis.ns1
	backend == backend.ns3
	redis-data == redis-data.ns2
namespace ns2
    www == www.ns2
	frontend == frontend.ns3
	backend == backend.ns3
    redis == redis.ns2
	backend == backend.ns3
	redis-data == redis-data.ns2
    redis-data == redis-data.ns2
namespace ns3
    frontend == frontend.ns3
    backend == backend.ns3
```

Above, we show an isolated ns1 using resources from ns2 and ns3 to
actually run. ns1 and ns2 are sharing the networks declared in ns3.

Please check out the simulation code for details but all references are
actually calculated from the perspective of the namespace. We can
actually make it such that changes to the namespace can allow one to
update entire applications with small changes to the namespace.

Signed-off-by: Stephen J Day <stephen.day@docker.com>