events: add events package
We take the events package from docker/distribution remove some
functionality, fix some consistency and performance issues. The package
provides a solid foundation for the efficient, goroutine safe
distribution of events.

The largest change from distribution is that only a single event is
dispatched at a time. Practice has shown this not to be useful. If
applications require it, they can define and envelope.

This package also omits the peicemeal metrics from distribution. These
will be replaced with an integrated solution to be compatible with
whatever we need.

Signed-off-by: Stephen J Day <stephen.day@docker.com>