Raft ID stability
Save raft node ID into WAL. When starting with an existing WAL, restore
the node ID and rejoin the cluster under the same ID as before.

Change NewNode not to take an ID, since the ID is either loaded from
disk or generated internally. This generation should be moved to the
remote side eventually.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>