Add a custom Picker that forces a reconnect attempt if necessary before each request
This works around GRPC's exponential backoff, which causes problems when
nodes rejoin. The leader's connection to a node that was once part of
the cluster but has been down for awhile could be sitting for a long
time between retries, and the join attempt would go wrong when the
rejoining node concludes that the leader is unreachable.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>