Include cluster member list and removed nodes in snapshots
Normally, the member list is built from configuration changes in the
raft log. When restarting a node, it will replay the raft log and build
the member list. However, if a snapshot has been taken and the log was
compacted, configuration change entries may no longer be present in the
log. This means that the snapshot needs to contain the member list (and
list of removed nodes) at the time the snapshot was taken.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>