Let go handle build deps
We shouldn't try to figure out go build dependencies in Make, because
we won't be able to do it as accurately as go. Add the binaries targets
to .PHONY so they are run without regard to dependencies. Then let go
figure out for itself whether anything needs to be rebuilt.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>