api: namespace API sketch
The required methods for adding namespace support are now present. We
define the first-class `Namespace` object which can be created, listed,
and deleted like any other. We also define the entries on other objects
that have namespace membership.

Once we confirm this API design, we can start work on implementing the
behavior within the datastore and from a CLI level.

Signed-off-by: Stephen J Day <stephen.day@docker.com>