Ignore --join-cluster when cluster state already exists
Previously, when --join-cluster was specified but couldn't be acted on
because there was already cluster state, we would log a warning but
leave the manager in an unusable state. This commit changes the behavior
to still log that warning, but continue starting up as a member of the
preexisting cluster.

Also, change the calling conventions of raft.go to make it more
consistent with other subsystems such as the scheduler. Make the address
of the cluster to join one of the new node options, instead of passing
it directly to StartByJoining. Unify Start and StartByJoining under the
name Run, and make it a blocking function like other subsystems' Run.
Instead of returning an error channel from Run, which could lead to
deadlocks without a dedicated monitor goroutine, log non-fatal errors
directly.

Fixes #240

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>