raft: Add a test covering cluster member downtime during a snapshot operation
Add TestSnapshotRestart, which stops one of the cluster members before
triggering a snapshot. During the downtime window, new members are added
to the cluster and new data is added to the store, both before and after
the snapshot happens. The restarted member must receive this snapshot
and get caught up with the state of the majority.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>