store: Don't check version number when applying an update from raft
When Update is called directly, the Version field contains the old
version, and we update it as part of the update process. But when we are
applying an update from the raft log, the version number has already
been changed to contain the new version, and comparing it against the
old version will fail. When we're operating in a passive mode (as a
follower, or as a leader replaying the log), don't do the version check,
and just apply the Version field as is. This makes sense because then
the sequencer is something that should only run on the leader.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>