Add networking support in agent
This commit adds networking support in agent by processing
task downloads and detecting network references in the task.
When a network reference is detected, a network is created
on the first reference and removed when the last task with a
reference to that network is DEAD.

Also, injected task side networking configuration so that
containers are properly plumbed into the appropriate network.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>