all: update code to use new API layout
We now update the code to use the new schema layout.

This PR indicates that we are likely way over coupled to the schema.
Ideally, we shouldn't touch fairly unrelated files. Part of this
involved separating a number of unrelated types but it is clear that we
are not abstracting operations.

Signed-off-by: Stephen J Day <stephen.day@docker.com>