api: add timestamp to task status message
Adds the timestamp type to TaskStatus. This can open up some interesting
behavior in the scheduler. For now, we let the agent set it, which may
not be ideal.

Note that we've had to do some nasty protobuf wrangling. External types
still aren't well supported. We'll be able to remove all of this once
gogoprotobuf supports timestamp.

Signed-off-by: Stephen J Day <stephen.day@docker.com>