api: provide the concept of service mode
Service mode defines the reconcialtion strategy for the tasks of a given
service. The default mode, `RUNNING`, instructs the orchstrator to
ensure that there are `instances` running tasks in the cluster. The mode
`BATCH` ensures that there are at least `instances` tasks that are in
the completed state.

These definitions help to form the model for global service scheduling.

Signed-off-by: Stephen J Day <stephen.day@docker.com>