Use configrable GRPC backoff for raft
The picker nonsense is no longer necessary now that we can limit the
maximum exponential backoff delay to a few seconds. Behavior seems to be
similar to what we had with the picker. Retries occur every few seconds,
and killing a manager and restarting it causes it to rejoin the cluster
quickly.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>