Introduce restart policy, desired state, and terminal state
This commit adds a restart policy. For now, it's a simple policy with
the following options:

- Always restart
- Restart on failure
- Never restart

The default is "always restart".

There is also a field for a restart delay. This is not implemented yet.

To support the restart policy, the orchestrator and drainer no longer
delete tasks. Instead they set a new "desired state" field to DEAD. This
tells the agent that it should kill the task if it's still running.

The orchestrator also needs to know whether a task failed or completed
successfully. This uses a new "terminal state" field in status that is
set by the agent. The orchestrator recognizes a "terminal state" of
COMPLETED as a successful task. A failed task is identified either
through a desired state of DEAD (if it was failed by the manager, for
example if it did not receive an agent heartbeat), or by a terminal
state of REJECTED or FAILED (if it was rejected/failed by the agent).

For now, completed tasks will never be deleted from the store. Task
cleanup will come in a future PR.

Currently, the orchestrator has no way of distinguishing a task that was
stopped because of scaling down a service from one that actually failed.
This means that when the restart policy is "restart on success" or
"restart never", tasks that were previously killed by scaling down will
count towards the instance count. For example, If you scale down a
service from 10 instances to 5, and then scale it back to 10, the 5
instances that were previously killed will count as failed instances,
and the orchestrator won't start any new ones. This will probably be
fixed when the orchestrator is revised to do this counting in another
way from listing tasks in the store (once tasks get garbage collected).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>