log: standardize on local logging package
This commit removes direct references to logrus throughout the code
base. We also ensure that the logger is grabbed from a local context,
where available.

The context has been plumbed where missing but full functionality has
not been added because certain cases require further analysis. The is
especially true in raft where context cancellation could be a critical
behavioral tool. The context.TODO() placeholder is used in cases where
the intended context is unclear or undefined.

Signed-off-by: Stephen J Day <stephen.day@docker.com>