api: add space between task state values
The state of a task is actually a lamport clock, in that
given two observations, the greater of the two can be considered
correct. To enforce this, we only allow tasks to proceed to a greater
state.

A byproduct of this design decision is that we must also maintain this
invariant in the protobuf enum values, such that when comparing two
values, the one with the greater value is also the greater state.

Because we may want to add intervening states a later date, we've left
64 spaces between each one. This should allow us to make 5 or 6
insertions between each state if we find that we made a mistake and need
another state.

Signed-off-by: Stephen J Day <stephen.day@docker.com>