Fix scheduler panic
A common pattern in the scheduler is the following:

        nodeInfo := s.nodeHeap.nodeInfo(t.NodeID)
        nodeInfo.removeTask(t)
        s.nodeHeap.updateNode(nodeInfo)

We explicitly don't check if the node exists, to keep the code simple.
It's assumed that updateNode will drop the update if we're dealing with
a nonexistent node.

However, updateNode was missing a check for a nil nodeInfo.Node. Add
this to updateNode and addOrUpdateNode.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>