Disable errcheck Makefile target
It's annoying me that running make runs golint twice, once as part of
the lint target and a second time under the bogus errcheck target.

The codebase is not errcheck-clean right now (see #272), so this is not
as simple as fixing the target to actually run errcheck.

I am disabling the errcheck target for now. If someone goes through the
effort to make the codebase pass errcheck, we can reenable it. But for
now, we're better off without a misleadingly named target that actually
runs golint a second time.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>