Fix removal of tasks in agent
We were checking if the desired state of the task was DEAD before
removing it, but this check was done against the original version of the
task assigned to the agent, not the updated assignment set.

Since this check is only useful for stopping a task without removing the
container, which isn't supported yet, just remove it for now.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>