raft: swarmctl managers ls to list raft managers in the cluster
This PR adds a new command, 'swarm managers ls' to list the set
of Managers in the cluster. This shows basic informations with
the ID, the Address and the Status of the node, reachable or
unreachable.

The PR also extracts some utility functions of raft_test.go in
its own package 'testutils' for other packages to have access to
those methods.

Signed-off-by: Alexandre Beslic <alexandre.beslic@gmail.com>