Combine orchestrator and drainer
The previous separation between the orchestrator and drainer caused some
extra complexity because both the orchestrator and the drainer could
kill tasks by setting their "desired state" to DEAD. The orchestrator
had to observe these transitions and restart tasks when appropriate.
This involved running a costly reconcillation function after each
change, because it couldn't know whether an updated task was already
dead, or if it had just been killed and needed to be replaced.

With the drainer merged into the orchestrator, we can separate
per-task reconcillation with per-service reconcillation. The per-service
reconcillation is the traditional loop, which happens whenever a service
is created or updated. When tasks or nodes die, we can kick off
replacements for those particular tasks without having to consider all
the other tasks attached to the service.

This also fixes a problem with the recursive nature of previous
orchestrator. It would create tasks, which would trigger reconcillation
again. This works correctly when the orchestrator runs in an atomic
transaction (though the additional reconcillation loop makes things
slower than necessary), but as of the batch size limiting, the
orchestrator no longer creates all its tasks atomically.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>