raft: Cancel contexts after operation completes
The send function creates a context with a 10 second timeout and never
cancels it. This means resources associated with the context will be kept
around for a full 10 seconds, as well as an unneeded timer.

While this probably won't break anything, it's better to cancel the
context once it has served its purpose. Have each sendToMember goroutine
create its own derived context and cancel it once done.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>