Add instance numbers to tasks
For tasks associated with "running" services, include an instance number
in the task object. This is the first step towards enabling users to
follow the history of a particular instance - i.e. a task was scheduled
on node A, node A failed, and then a replacement task was scheduled on
node B. It will also allow task history pruning, since we'll be able to
guarantee that we keep a certain number of the most recent task objects
for each instance.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>