Add limit on the size of task history
So far, the orchestrator never deletes tasks. It only sets the "desired
state" of defunct tasks to DEAD. Over time, these dead tasks could
occupy a lot of storage, and make the UI difficult to interact with.

This commit adds a limit on how many old tasks to track.  When this
limit is reached, the oldest non-running tasks are deleted.

For running jobs, the limit is imposed per instance number. Each
instance of a task associated with a "running" service will have the
currently executing task, if any, and older tasks preceding this one, up
to the limit.

For fill jobs, the limit is imposed per node.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>