manager: Create listener earlier
When joining a raft cluster, a manager creates a raft node before
creating its GRPC listener. The raft node was invoking an RPC call to
the address used to join the cluster, and the host on the other end
would try to open a connection back. This didn't always work on the
first try because the manager did not create its GRPC listener until
later. This could cause messages like this:

2016/04/26 16:10:54 grpc: Conn.resetTransport failed to create client
transport: connection error: desc = "transport: dial tcp 0.0.0.0:4243:
getsockopt: connection refused"; Reconnecting to "0.0.0.0:4243"

To solve this, create the listener before the raft node.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>