raft: Make test send timeouts looser
The tests were using 100 ms timeouts for sends to other cluster members,
which could potentially lead to flaky behavior in an oversubscribed
virtualized environment. Boost the timeouts to 10 seconds. To prevent
this from adding to the time it takes to tear down the cluster after
each test, cancel the raft node's parent context when it shuts down, so
sends are also cancelled.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>