api: consolidate protobuf types into a single package
Enabled by making deepcopy more granular, we have merged all of the
protobuf types into a single package. Not only that, it only takes a
single pass to generate all related protobuf files for the main API
package.

While this seems small, it will make it easier to split up in the
future.

Signed-off-by: Stephen J Day <stephen.day@docker.com>