store: View doesn't need to return an error
Viewing the store can't actually fail. The current View method returns
an error which is just a passthrough of the error returned by the
callback. This is convenient for situations that want to pass an error
through, but for most situations, it's just an extra "return nil" in the
callback and an extra error check after View.

This changes View and its callback type not to return errors. In the
cases where a callback needs to signal an error, it can do so by setting
a variable that's visible outside the closure.

Diffstat is:

    20 files changed, 170 insertions(+), 261 deletions(-)

...so I would say this is a net simpification.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>