Stop allocator when manager shuts down
This missing call could lead to a panic on shutdown when the running
allocator called into the stopped raft store.

Also, convert allocator to use stopChan/doneChan pattern like the rest
of the services, since the race detector was complaining about
cancelling the context from a different goroutine than the one that
created it once the call to Stop was added.

Fixes #457

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>