scheduler: Handle store errors more gracefully
If some task updates fail, don't give up right away. Keep track of the
failed updates, but continue to make the rest of the updates in the
batch. Roll back local state for the failed updates, and also anything
that store.Batch indicates was not applied.

Combine the task update loops from tick and processPreassignedTasks into
a common function to reduce code duplication.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>