agent: remove statuses map
Before, we maintained a separate map for tasks and their statuses. We
now just maintain the two in a single map. We need to be careful not to
overwrite status in certain cases, but this provides a simpler model
that should use less memory.

This also fixes a bug where status was being properly updated after a
switch away from pointers for TaskStatus.

Signed-off-by: Stephen J Day <stephen.day@docker.com>