Remove store interface
The Store interface originally had methods such as Nodes, Tasks, etc.
that returned interfaces with typed methods to operate on that specific
object type. This was really nice for the consumer of the store API, but
it was painful to implement, because the store implementation had to
know about all the different types it could support (and implement an
interface for each).

This removes the Store interface and replaces those methods with typed
helpers such as CreateNode, DeleteNode, FindNodes, etc. This means the
implementation of an object type is completely separate from the memory
store core, and can be entirely contained in its own file.

Also move some store-specific variables and types from state/store.go to
the store package, now that they are not needed to support the interface
definition.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>