manager: Fix racy service starting
The manager was starting goroutines to run the scheduler, orchestrator,
etc, and those goroutines would use variables like m.scheduler and
m.orchestrator without holding a lock. This could potentially
dereference a nil pointer, or even call Run on the same item twice if
the node stopped being the leader for a very short period.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>