raft: shutdown raft node and manager components after removal
This commit fixes the removal process to initiate the shutdown
logic after a node has been removed by other members and tries
to contact them back. This circumvents the issue of a node not
processing the CC and having to 'guess' whether it has been
removed or not.

This also fixes issues with members trying to mess up with the
election after they've been removed from the memberlist. Now
ProcessRaftMessage calls are returned with an error if a node
tries to send a message to other members when its ID is found
in the remove set.

Signed-off-by: Alexandre Beslic <alexandre.beslic@gmail.com>