Accept only NetworkID for NetworkAttachment config
In Container, currently we are allowing either a name or a network ID to
refer to the network attachment. This is unnecessary. It is much simpler
to always refer the networks in network attachments using an ID. That
way even if a network name changes in the future we don't need to update
all the references. Client side code may use network names and resolve
it client side to an ID before creating or updating the object.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>