agent: better handling of state resolution
Agent task state reporting is slightly cleaned up here for tasks that
are already running. The most interesting change happens directly in the
controller where we handle exited containers on the Wait call. Task
updates are also de-bounced in this changeset, removing a lot of the
warning messages encountered in the controller.

Signed-off-by: Stephen J Day <stephen.day@docker.com>