agent: label containers with service annotations
The main change here is moving to "nice" container names. Much of the
cruft present before has been removed in favor of just the task name and
id. To allow querying by various aspects, we've included the node id,
the service id and task id labels.

We also have renamed annotations on Task to ServiceAnnotations to better
reflect the source of the annotations. We may want to have containers be
explicitly labelled, so that option is left open.

Signed-off-by: Stephen J Day <stephen.day@docker.com>