Add limit on number of restarts
Add MaxAttempts and Window fields to RestartPolicy.  MaxAttempts is the
number of restart attempts before giving up, if nonzero. Window is the
period to look back when enforcing the restart policy. If it is zero, the
lookback is indefinite.

Enforce these parameters in RestartSupervisor, using local state. Add
unit tests.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>