dispatcher: Batch task updates
The dispatcher currently does a raft write for each task status it needs
to update. This can become a bottleneck because a raft write can involve
significant latency.

This commit changes the dispatcher to maintain a map of outstanding
updates. When this map reaches a certain size, or a certain amount of
time elapses, the updates are flushed to raft in a batch.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>